Revert "Added Administrator and Signature Model Modified Holder Model to include List of Administrators Moved Jeph created Administrator to org.greenpole.entity.model"
This reverts commit c166bb67e0fa0aa94d647add13ec7318318f59d4.
